# contracts/openapi/public-discovery-api-v1.yaml
openapi: 3.0.3
info:
  title: Public Discovery & Insights API
  version: 1.0.0
  description: Public API for Discovery & Insights
servers:
  - url: https://api.example.com
    description: Production server
paths:
  /api/v1/stations:
    get:
      operationId: listStations
      summary: List charging stations
      parameters:
        - name: latitude
          in: query
          schema:
            type: number
        - name: longitude
          in: query
          schema:
            type: number
        - name: radius
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of stations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StationSummary'
  /api/v1/stations/{stationRef}:
    get:
      operationId: getStationDetails
      summary: Get public station details
      parameters:
        - name: stationRef
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Station details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationDetails'
              example:
                ref: "SEEDSTA0001"
                name: "Seed Fixture — Athens Center"
                address: "Fixture Street 1"
                latitude: 37.983810
                longitude: 23.727540
                city: "Athens"
                postalCode: "10431"
                countryCode: "GR"
                updatedAt: "2026-09-12T00:00:00Z"
                sourceVersion: 0
        '404':
          description: Station not found
          content:
            application/json:
              schema:
                $ref: ../schemas/common/problem-details.json
  /api/v1/stations/{stationRef}/availability:
    get:
      operationId: getStationAvailability
      summary: Get station availability
      parameters:
        - name: stationRef
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Station availability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationAvailability'
        '404':
          description: Station not found
          content:
            application/json:
              schema:
                $ref: ../schemas/common/problem-details.json
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
    anonymous:
      type: http
      scheme: bearer
      description: Anonymous access with rate limiting
  schemas:
    StationSummary:
      type: object
      properties:
        ref:
          type: string
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        availableEvses:
          type: integer
        totalEvses:
          type: integer
      required:
        - ref
        - name
        - latitude
        - longitude
    StationDetails:
      type: object
      properties:
        ref:
          type: string
          description: Public station reference
        name:
          type: string
          description: Station display name
        address:
          type: string
          description: Street address
        latitude:
          type: number
          description: Latitude in decimal degrees
        longitude:
          type: number
          description: Longitude in decimal degrees
        city:
          type: string
          description: City
        postalCode:
          type: string
          description: Postal code
        countryCode:
          type: string
          description: ISO 3166 alpha-2 country code
        updatedAt:
          type: string
          format: date-time
          description: Projection update timestamp (operational freshness)
        sourceVersion:
          type: integer
          description: Source aggregate version of the last applied event
      required:
        - ref
        - name
        - latitude
        - longitude
        - updatedAt
    StationAvailability:
      type: object
      properties:
        stationRef:
          type: string
        status:
          type: string
          enum:
            - available
            - busy
            - offline
            - limited
        availableEvses:
          type: integer
        totalEvses:
          type: integer
        nextAvailableSlot:
          type: string
          format: date-time
      required:
        - stationRef
        - status
        - availableEvses
        - totalEvses
security:
  - apiKey: []
  - anonymous: []